requirements.txt
appnope==0.1.0
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
cycler==0.10.0
dask==0.10.0
decorator==4.0.10
dill==0.2.5
functools32==3.2.3.post2
gnureadline==6.3.3
ipython==4.2.0
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
matplotlib==1.5.1
mpmath==0.19
networkx==1.11
numpy==1.11.0
pandas==0.18.1
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.2.0
ptyprocess==0.5.1
pyparsing==2.1.5
python-dateutil==2.5.3
pytz==2016.4
pyzmq==15.2.0
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.1
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sympy==1.0
Theano==0.8.2
toolz==0.8.0
tornado==4.3
traitlets==4.2.1
